extracting render controller into separate controlelrs
