+++ app/controllers/categories-controller.js
@@ -1,5 +1,9 @@
 'use strict';
 const Category = require('../models/category');
+const _ = require('lodash');
+const htmlencode = require('htmlencode').htmlEncode;
+const moment = require('moment');
+const numeral = require('numeral');
 
 class CategoriesController {
     static categories(req, res) {
