+++ app/controllers/pages-controller.js
@@ -1,4 +1,9 @@
 'use strict';
+const path = require('path');
+const _ = require('lodash');
+const htmlencode = require('htmlencode').htmlEncode;
+const moment = require('moment');
+const numeral = require('numeral');
 
 class PagesController {
     static join(req, res) {
@@ -11,6 +16,19 @@ class PagesController {
         res.locals.css = 'join-complete.css';
         res.locals.title = 'Thanks for joining the Open Data Network.';
         res.render('join-complete.ejs');
+    }
+
+    static error(req, res, statusCode, title) {
+        statusCode = statusCode || 500;
+        title = title || 'Internal server error';
+
+        return (error) => {
+            const type = statusCode < 500 ? 'client' : 'server';
+            const stack = error && error.stack ? `\n${error.stack}` : '';
+            console.error(`at=error status=${statusCode} type=${type} path="${req.path}" ${stack}`);
+            res.status(statusCode);
+            res.render('error.ejs', {statusCode, title});
+        };
     }
 
 }
