+++ app/controllers/search-controller.js
@@ -1,25 +1,45 @@
 'use strict';
 const API = require('./api');
-const Questions = require('./questions');
+const Questions = require('../../controllers/questions');
+const Data = require('../../controllers/data');
+const Navigate = require('../../controllers/navigate');
+const Constants = require('../../controllers/constants');
+const Relatives = require('../../controllers/relatives');
+
 const SrcConstants = require('../src/constants');
+const MapSources = require('../src/data/map-sources');
+const MapDescription = require('../src/maps/description');
+
 const EntityFormatter = require('../lib/entity-formatter');
 const SearchHelper = require('../lib/search-helper');
 const DatasetHelper = require('../lib/dataset-helper');
 const DataHelper = require('../lib/data-helper');
+const ParamsHelper = require('../lib/params-helper');
+const PagesController = require('./pages-controller');
+
+const DatasetConfig = require('../src/dataset-config');
+const quickLinksCount = 15;
+const refineByCount = 5;
+
+const defaultMetaSummary = 'Find the data you need to power your business, app, or analysis from across the open data ecosystem.';
+const _ = require('lodash');
+const htmlencode = require('htmlencode').htmlEncode;
+const moment = require('moment');
+const numeral = require('numeral');
 
 class SearchController {
     static search(req, res) {
-        RenderController._parameters(req, res).then(params => {
+        ParamsHelper.parameters(req, res).then(params => {
             try {
                 SearchController._search(req, res, params);
             } catch (error) {
-                RenderController.error(req, res)(error);
+                PagesController.error(req, res)(error);
             }
-        }, RenderController.error(req, res));
+        }, PagesController.error(req, res));
     }
 
     static searchResults(req, res) {
-        RenderController._parameters(req, res).then(params => {
+        ParamsHelper.parameters(req, res).then(params => {
             API.datasets(params).then(searchResults => {
                 if (searchResults.results.length === 0) {
                     res.status(204);
@@ -35,11 +55,11 @@ class SearchController {
 
                         res.render('_search-results-items.ejs', templateParams);
                     } catch (error) {
-                        RenderController.error(req, res)(error);
+                        PagesController.error(req, res)(error);
                     }
                 }
-            }, RenderController.error(req, res));
-        }, RenderController.error(req, res));
+            }, PagesController.error(req, res));
+        }, PagesController.error(req, res));
     }
 
     static searchWithVector(req, res) {
@@ -60,7 +80,7 @@ class SearchController {
             return false;
         }
 
-        RenderController._parameters(req, res).then(params => {
+        ParamsHelper.parameters(req, res).then(params => {
 
             if ((vector === '') || !containsDataset(params.dataAvailability, vector)) {
                 toDefaultVector();
@@ -70,9 +90,9 @@ class SearchController {
             try {
                 RenderController._search(req, res, params);
             } catch (error) {
-                RenderController.error(req, res)(error);
+                PagesController.error(req, res)(error);
             }
-        }, RenderController.error(req, res));
+        }, PagesController.error(req, res));
     }
 
     static _search(req, res, params) {
@@ -179,9 +199,9 @@ class SearchController {
 
                     res.render('search.ejs', templateParams);
                 } catch (error) {
-                    RenderController.error(req, res)(error);
+                    PagesController.error(req, res)(error);
                 }
-            }, RenderController.error(req, res));
+            }, PagesController.error(req, res));
         }
     }
 
@@ -387,12 +407,12 @@ class SearchController {
 
                     res.render('search.ejs', templateParams);
 
-                }, RenderController.error(req, res));
+                }, PagesController.error(req, res));
 
             } catch (error) {
-                RenderController.error(req, res)(error);
+                PagesController.error(req, res)(error);
             }
-        }, RenderController.error(req, res));
+        }, PagesController.error(req, res));
     }
 
 }
